backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
yfinance==0.2.40
ccxt==4.3.0
ta==0.11.0
langchain==0.3.0
langchain-openai==0.2.0
langchain-anthropic==0.2.0
langchain-groq==0.2.0
langchain-google-genai==2.0.0
transformers==4.44.0
torch==2.4.0
xgboost==2.1.0
prophet==1.1.5
pandas==2.2.0
numpy==1.26.0
sqlalchemy==2.0.0
aiosqlite==0.20.0
python-dotenv==1.0.0
httpx==0.27.0
finnhub-python==2.4.27
pydantic==2.9.0
